refactor: move transcript playback scenarios
