+++ src/loushang/coding/ui/playback_runner.py
@@ -7,20 +7,12 @@ from collections.abc import Sequence
 from pathlib import Path
 from typing import TextIO
 
-from loushang.coding.ui.perf_probe import build_synthetic_long_transcript_records
-from loushang.coding.ui.playback import (
-    NativeTuiInputPlaybackResult,
-    NativeTuiInputScenario,
-)
-from loushang.coding.ui.playback_scenarios.budgets import (
-    INTERACTION_FRAME_BUDGET,
-    LONG_TRANSCRIPT_FRAME_BUDGET,
-)
 from loushang.coding.ui.playback_scenarios.command import COMMAND_ROUTING_SCENARIOS
 from loushang.coding.ui.playback_scenarios.composer import COMPOSER_SCENARIOS
 from loushang.coding.ui.playback_scenarios.lifecycle import LIFECYCLE_SCENARIOS
 from loushang.coding.ui.playback_scenarios.surface import SURFACE_SCENARIOS
 from loushang.coding.ui.playback_scenarios.terminal import TERMINAL_SCENARIOS
+from loushang.coding.ui.playback_scenarios.transcript import TRANSCRIPT_SCENARIOS
 from loushang.coding.ui.playback_suite import (
     NativePlaybackScenarioResult,
     NativePlaybackScenarioSpec,
@@ -29,7 +21,15 @@ from loushang.coding.ui.playback_suite import (
 from loushang.coding.ui.playback_suite import (
     run_playback_scenarios as _run_playback_scenarios,
 )
-from loushang.tui.transcript import ToolExecutionRecord
+
+__all__ = [
+    "DEFAULT_SUITE",
+    "NativePlaybackScenarioResult",
+    "NativePlaybackScenarioSpec",
+    "NativePlaybackSuite",
+    "run_playback_cli",
+    "run_playback_scenarios",
+]
 
 
 def run_playback_scenarios(
@@ -132,68 +132,13 @@ def _json_summary(results: Sequence[NativePlaybackScenarioResult]) -> dict[str, 
     }
 
 
-def _run_long_transcript_input() -> NativeTuiInputPlaybackResult:
-    result = (
-        NativeTuiInputScenario(width=100, height=18)
-        .with_records(build_synthetic_long_transcript_records(turns=40, tail_tool_output_lines=300))
-        .render()
-        .type_chars("fresh input")
-        .run()
-    )
-    result.assert_composer_text("fresh input")
-    result.assert_visible_contains("› fresh input")
-    result.assert_no_clear_screen()
-    LONG_TRANSCRIPT_FRAME_BUDGET.assert_result(result, skip_first=True)
-    result.assert_screen_anchor_stable("›", occurrence="last")
-    return result
-
-
-def _run_tool_output_preview() -> NativeTuiInputPlaybackResult:
-    result = (
-        NativeTuiInputScenario(width=100, height=16)
-        .with_records(
-            (
-                ToolExecutionRecord(
-                    name="bash pytest tests/coding -q",
-                    state="completed",
-                    elapsed_seconds=0.6,
-                    output="\n".join(f"line {index}" for index in range(1, 13)),
-                ),
-            )
-        )
-        .render()
-        .type_text("next")
-        .run()
-    )
-    result.assert_visible_contains("  └ line 1")
-    result.assert_visible_contains("    line 3")
-    result.assert_visible_contains("    ... (6 hidden lines)")
-    result.assert_visible_contains("    line 12")
-    result.assert_visible_not_contains("    line 4")
-    result.assert_visible_not_contains("    line 9")
-    result.assert_visible_contains("› next")
-    result.assert_no_clear_screen()
-    result.assert_cursor_matches_diagnostics()
-    INTERACTION_FRAME_BUDGET.assert_result(result, skip_first=True)
-    return result
-
-
 DEFAULT_SUITE = NativePlaybackSuite(
     (
         *COMPOSER_SCENARIOS,
         *LIFECYCLE_SCENARIOS,
         *COMMAND_ROUTING_SCENARIOS,
         *SURFACE_SCENARIOS,
-        NativePlaybackScenarioSpec(
-            name="long-transcript-input",
-            description="Echo input after a long transcript using bounded frame updates.",
-            run=_run_long_transcript_input,
-        ),
-        NativePlaybackScenarioSpec(
-            name="tool-output-preview",
-            description="Render long tool output as head, hidden-count, and tail without flicker.",
-            run=_run_tool_output_preview,
-        ),
+        *TRANSCRIPT_SCENARIOS,
         *TERMINAL_SCENARIOS,
     )
 )
